对数据库操作抽离出DAO,并抽象成接口
